Add client-side ML inference and fix audit bugs
Wire the trained Cd/Cl surrogate model into the browser so predictions
appear instantly as users adjust wind speed, Reynolds number, and model
selection. The forward pass runs in pure TypeScript with no ML framework
dependency (~525K params, <1ms per prediction).

Fix bugs found during codebase audit:
- Guard division by zero in coloring.c and camera normalize (main.c)
- Add malloc NULL check in opengl_utils.c shader loader
- NULL pointers after free in model_loader.c
- Remove debug printf from coloring hot paths
- Add epsilon guards to autodiff sqrt, log, exp, reciprocal gradients
- Fix AdamW epsilon placement: sqrt(v_hat) + eps -> sqrt(v_hat + eps)
- Use size_t in tensor numel computation to prevent overflow
- Abort training on NaN/Inf loss
- Fix parseHash NaN passthrough from malformed URL parameters
- Add timeout to Modal render POST request
- Gitignore website model weight copies

+++ ml/framework/src/tensor.cpp
@@ -9,9 +9,9 @@
 
 int Tensor::compute_numel(const std::vector<int>& shape) {
     if (shape.empty()) return 0;
-    int n = 1;
-    for (int d : shape) n *= d;
-    return n;
+    size_t n = 1;
+    for (int d : shape) n *= (size_t)d;
+    return (int)n;
 }
 
 void Tensor::sync_rows_cols() {
